requirements.txt
pillow==10.2.0
pygame==2.5.2
requests==2.31.0
pyfiglet==0.8.post1
colorama==0.4.6 